llama_3_2_1b: flowkv -> LUT-exp + fp32 probs + exact div, BIT-EXACT
Replace v0's HW transcendentals + bf16 stream with bit-exact-friendly
equivalents:

  - LUT-based exp (was aie::exp2): kernel and reference index the
    SAME precomputed 256-entry fp32 exp LUT generated by
    gen_exp_lut.py. After subtract-max, shifted scores are quantized
    to int8 with EXP_QUANT_SCALE (default 0.05); both kernel and
    numpy reference use identical quant arithmetic + identical LUT
    bytes -> exp value is byte-identical.

  - Exact 1.0f / sum (was aie::inv): HW reciprocal differs from
    fp32 division at the last bit; swap for exact division.

  - fp32 probs on the CT0 -> CT1 stream (was bf16): eliminates the
    bf16 truncation on softmax probabilities between qk and sv.
    T*4 = 64 bytes for T=16 -- negligible bandwidth cost vs bit-
    exact gain.

  - Single combined qk_scale multiply (was two separate fp32 muls):
    matches numpy reference's combined-scale ordering so fp rounding
    is deterministic.

Validation at head_dim=64, T=16:
  mismatches=0/64  max|int8 diff|=0  BIT-EXACT PASS

Pattern follows yolo_m10_softmax (LUT keyed by int8 of SRS'd scores).
gen_exp_lut.py emits a uint32-bits header included by the kernel and
re-computed in numpy with identical fp32 arithmetic.

Follow-ups: vec-ize the inner loops (scalar in v0 to keep order
deterministic); the LUT-keyed pattern can be vectorized too.

+++ programming_examples/ml/llama_3_2_1b/aie2_flowkv.py
@@ -20,7 +20,6 @@ import argparse
 import sys
 
 import numpy as np
-from ml_dtypes import bfloat16
 
 from aie.iron import Kernel, ObjectFifo, Program, Runtime, Worker
 from aie.iron.device import NPU2, Tile
@@ -40,7 +39,9 @@ INV_OUT_SCALE = 1.0 / 0.05
 def build(head_dim: int, t: int):
     q_ty     = np.ndarray[(head_dim,),       np.dtype[np.int8]]
     kv_ty    = np.ndarray[(t * head_dim,),   np.dtype[np.int8]]
-    probs_ty = np.ndarray[(t,),              np.dtype[bfloat16]]
+    # fp32 probs on the CT0 -> CT1 stream; no bf16 truncation so the
+    # softmax probabilities arrive bit-identical to what qk computed.
+    probs_ty = np.ndarray[(t,),              np.dtype[np.float32]]
     out_ty   = np.ndarray[(head_dim,),       np.dtype[np.int8]]
 
     of_q     = ObjectFifo(q_ty,     name="q")
